build out more apiProxy tests
